Moves classes into enemy packages
