+++ src/main/java/andioopp/domain/model/enemy/enemies/MontyMole.java
@@ -1,0 +1,10 @@
+package andioopp.domain.model.enemy.enemies;
+
+import andioopp.common.transform.Transform;
+import andioopp.domain.model.enemy.Enemy;
+
+public class MontyMole extends Enemy {
+    public MontyMole(Transform transform, String sprite) {
+        super(transform, sprite);
+    }
+}
