+++ src/main/java/andioopp/domain/model/towers/Luigi.java
@@ -1,0 +1,4 @@
+package andioopp.domain.model.towers;
+
+public class Luigi extends Tower{
+}
