+++ src/main/java/andioopp/domain/model/towers/Mario.java
@@ -1,0 +1,4 @@
+package andioopp.domain.model.towers;
+
+public class Mario extends Tower{
+}
