+++ src/main/java/andioopp/domain/model/towers/Toad.java
@@ -1,0 +1,4 @@
+package andioopp.domain.model.towers;
+
+public class Toad extends Tower{
+}
