+++ src/main/java/andioopp/domain/model/towers/Yoshi.java
@@ -1,0 +1,4 @@
+package andioopp.domain.model.towers;
+
+public class Yoshi extends Tower{
+}
