--- src/main/java/andioopp/model/enemies/Blooper.java
@@ -1,4 +1,0 @@
-package andioopp.model.enemies;
-
-public class Blooper{
-}
