--- src/main/java/andioopp/model/enemies/Bowser.java
@@ -1,4 +1,0 @@
-package andioopp.model.enemies;
-
-public class Bowser {
-}
