--- src/main/java/andioopp/model/enemies/CheepCheep.java
@@ -1,6 +1,0 @@
-package andioopp.model.enemies;
-
-public class CheepCheep {
-
-
-}
