--- src/main/java/andioopp/model/enemies/HammerBro.java
@@ -1,10 +1,0 @@
-package andioopp.model.enemies;
-
-import andioopp.common.Transform;
-import andioopp.gfx.Sprite;
-
-public class HammerBro extends Enemy {
-    public HammerBro(Transform transform, Sprite sprite) {
-        super(transform, sprite);
-    }
-}
