--- src/main/java/andioopp/model/enemies/MontyMole.java
@@ -1,10 +1,0 @@
-package andioopp.model.enemies;
-
-import andioopp.common.Transform;
-import andioopp.gfx.Sprite;
-
-public class MontyMole extends Enemy{
-    public MontyMole(Transform transform, Sprite sprite) {
-        super(transform, sprite);
-    }
-}
