--- src/main/java/andioopp/model/towers/Bobomb.java
@@ -1,4 +1,0 @@
-package andioopp.model.towers;
-
-public class Bobomb {
-}
