--- src/main/java/andioopp/model/towers/Luigi.java
@@ -1,4 +1,0 @@
-package andioopp.model.towers;
-
-public class Luigi extends Tower{
-}
