--- src/main/java/andioopp/model/towers/Mario.java
@@ -1,4 +1,0 @@
-package andioopp.model.towers;
-
-public class Mario extends Tower{
-}
