--- src/main/java/andioopp/model/towers/Toad.java
@@ -1,4 +1,0 @@
-package andioopp.model.towers;
-
-public class Toad extends Tower{
-}
