--- src/main/java/andioopp/model/towers/Yoshi.java
@@ -1,4 +1,0 @@
-package andioopp.model.towers;
-
-public class Yoshi extends Tower{
-}
